@@ -4,13 +4,13 @@ import 'package:mp_chart/mp/core/highlight/highlight.dart';
 import 'package:mp_chart/mp/painter/painter.dart';
 
 abstract class HighlightUtils {
-  static Highlight performHighlight(
-      ChartPainter painter, Highlight curHighlight, Highlight lastHighlight) {
-    if (curHighlight == null || curHighlight.equalTo(lastHighlight)) {
-      painter.highlightValue6(null, true);
+  static Highlight? performHighlight(
+      ChartPainter? painter, Highlight? curHighlight, Highlight? lastHighlight) {
+    if (curHighlight == null || curHighlight.equalTo(lastHighlight!)) {
+      painter!.highlightValue6(null, true);
       lastHighlight = null;
     } else {
-      painter.highlightValue6(curHighlight, true);
+      painter!.highlightValue6(curHighlight, true);
       lastHighlight = curHighlight;
     }
     return lastHighlight;
@@ -18,7 +18,7 @@ abstract class HighlightUtils {
 
   static Highlight performHighlightDrag(
       ChartPainter painter, Highlight lastHighlight, Offset offset) {
-    Highlight h = painter.getHighlightByTouchPoint(offset.dx, offset.dy);
+    Highlight? h = painter.getHighlightByTouchPoint(offset.dx, offset.dy);
 
     if (h != null && !h.equalTo(lastHighlight)) {
       lastHighlight = h;
