@@ -98,13 +98,13 @@ abstract class Matrix4Utils {
     (Matrix4.identity()..setTranslationRaw(tx, ty, 0.0)).copyInto(m);
   }
 
-  static void mapPoints(Matrix4 m, List<double> valuePoints) {
-    double x = 0;
-    double y = 0;
+  static void mapPoints(Matrix4 m, List<double?> valuePoints) {
+    double? x = 0;
+    double? y = 0;
     for (int i = 0; i < valuePoints.length; i += 2) {
       x = valuePoints[i] == null ? 0 : valuePoints[i];
       y = valuePoints[i + 1] == null ? 0 : valuePoints[i + 1];
-      final Vector3 transformed = m.perspectiveTransform(Vector3(x, y, 0));
+      final Vector3 transformed = m.perspectiveTransform(Vector3(x!, y!, 0));
       valuePoints[i] = transformed.x;
       valuePoints[i + 1] = transformed.y;
     }
