@@ -4,8 +4,8 @@ import 'package:mp_chart/mp/core/utils/utils.dart';
 
 abstract class PainterUtils {
   static TextPainter create(
-      TextPainter painter, String text, Color color, double fontSize,
-      {String fontFamily, FontWeight fontWeight = FontWeight.w400}) {
+      TextPainter? painter, String? text, Color? color, double? fontSize,
+      {String? fontFamily, FontWeight? fontWeight = FontWeight.w400}) {
     if (painter == null) {
       return _create(text, color, fontSize,
           fontFamily: fontFamily, fontWeight: fontWeight);
@@ -13,9 +13,9 @@ abstract class PainterUtils {
 
     if (painter.text != null && (painter.text is TextSpan)) {
       var preText = (painter.text as TextSpan).text;
-      var preColor = painter.text.style.color;
+      var preColor = painter.text!.style!.color;
       preColor = preColor == null ? ColorUtils.BLACK : preColor;
-      var preFontSize = painter.text.style.fontSize;
+      var preFontSize = painter.text!.style!.fontSize;
       preFontSize =
           preFontSize == null ? Utils.convertDpToPixel(13) : preFontSize;
       return _create(
@@ -30,8 +30,8 @@ abstract class PainterUtils {
     }
   }
 
-  static TextPainter _create(String text, Color color, double fontSize,
-      {String fontFamily, FontWeight fontWeight = FontWeight.w400}) {
+  static TextPainter _create(String? text, Color? color, double? fontSize,
+      {String? fontFamily, FontWeight? fontWeight = FontWeight.w400}) {
     return TextPainter(
         textAlign: TextAlign.center,
         textDirection: TextDirection.ltr,
@@ -41,8 +41,8 @@ abstract class PainterUtils {
                 fontFamily: fontFamily, fontWeight: fontWeight)));
   }
 
-  static TextStyle createTextStyle(Color color, double fontSize,
-      {String fontFamily, FontWeight fontWeight = FontWeight.w400}) {
+  static TextStyle createTextStyle(Color? color, double? fontSize,
+      {String? fontFamily, FontWeight? fontWeight = FontWeight.w400}) {
     if (fontWeight == null) {
       fontWeight = FontWeight.w400;
     }
